pyproject.toml
[tool.poetry]
name = "sarl"
version = "0.1.1-alpha"
description = "Algorithms and environments for structured action-space reinforcement learning."
authors = ["Theodore Tollet <t.tollet@lancaster.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.9, <3.12"
numpy = "^1.24.4"
pygame = "^2.4.0"
numba = "^0.57.0"
gymnasium = "^0.28.1"
torch = "^1.13.1"
click = "^8.1.3"
tqdm = "^4.66.1"
hydra-core = "^1.3.2"
stable-baselines3 = "^2.3.0"
rich = "^13.7.1"
protobuf = "4.25.3"  # https://github.com/databricks/dbt-databricks/issues/614 encountered whilst using tensorboard
tensorboard = "^2.16.2"
ipykernel = "^6.29.5"
polars = "^1.21.0"
altair = "^5.5.0"
vl-convert-python = "^1.7.0"
vegafusion = {extras = ["embed"], version = "^2.0.1"}
vegafusion-python-embed = "^1.6.9"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.11.0"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Addition here enables running without using python3 preface
# entrypoints, etc []
# sarl.train:main

# [project.scripts]
# train = "sarl.train:main"
